# webserver/swagger.yml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: ML Aide
  license:
    name: unspecified
servers:
  - url: http://api.mlaide.io/v1
security:
  - oauth2: []
paths:
...
  /projects/{projectKey}/runs:
    get:
      summary: List all runs of a project
      operationId: listRuns
      tags:
        - runs
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the experiments belong to
          schema:
            type: string
        - name: runKeys
          in: query
          required: false
          description: Keys of runs that shall be returned
          schema:
            type: array
            items:
              type: number
        - name: experimentKey
          in: query
          required: false
          description: Key of experiment that returned runs shall contain as experimentRef
          schema:
            type: string
      responses:
        '200':
          description: An object containing an array of runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runs"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new run
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Run"
      tags:
        - runs
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the run should belong to
          schema:
            type: string
      responses:
        '200':
          description: Run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{projectKey}/runs/{runKey}:
    get:
      summary: Get single run by runKey
      operationId: getRunByKey
      description: Get single run by runKey
      tags:
        - runs
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the run belongs to
          schema:
            type: string
        - name: runKey
          in: path
          required: true
          description: The key of the run that should be updated
          schema:
            type: number
      responses:
        '200':
          description: An object containing the requested run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Partial update an existing run
      operationId: partialUpdateRun
      description: "Update an existing run. This operation executes a partial update.
        That means, that only properties that should be modified must be contained in the request body.
        Only the status, parameters and metrics can be updated. If you update parameters or metrics with this
        operation, all existing parameters/metrics are replaced. If you want to add a parameter/metric to the
        existing ones, you should use PUT /projects/{projectKey}/runs/{runKey}/parameters or
        PUT /projects/{projectKey}/runs/{runKey}/metrics"
      tags:
        - runs
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the run belongs to
          schema:
            type: string
        - name: runKey
          in: path
          required: true
          description: The key of the run that should be updated
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/Run"
      responses:
        '204':
          description: The update was successful
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{projectKey}/runs/{runKey}/note:
    put:
      summary: Create a new or update the existing note for an existing run
      operationId: createOrUpdateRunNote
      tags:
        - runs
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the run belongs to
          schema:
            type: string
        - name: runKey
          in: path
          required: true
          description: The key of the run that should be updated
          schema:
            type: number
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: This was an awesome run!
      responses:
        '200':
          description: The updated note
          content:
            text/plain:
              schema:
                type: string
                example: This was an awesome run!
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{projectKey}/runs/{runKey}/parameters:
    patch:
      summary: Merge/Patch a parameter object into the existing parameter object
      operationId: updateRunParameters
      tags:
        - runs
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the run belongs to
          schema:
            type: string
        - name: runKey
          in: path
          required: true
          description: The key of the run that should be updated
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/RunParameter"
      responses:
        '204':
          description: The operation was successful
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /projects/{projectKey}/experiments:
    get:
      summary: List all experiments of a project
      operationId: listExperiments
      tags:
        - experiments
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the experiments belong to
          schema:
            type: string
      responses:
        '200':
          description: An object containing an array of experiments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiments"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new experiment
      operationId: createExperiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Experiment"
      tags:
        - experiments
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the experiment should belong to
          schema:
            type: string
      responses:
        '200':
          description: Experiment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /projects/{projectKey}/experiments/{experimentKey}:
    get:
      summary: Get single experiment by experimentKey
      operationId: getExperimentByKey
      description: Get single experiment by experimentKey
      tags:
        - experiments
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the experiment belongs to
          schema:
            type: string
        - name: experimentKey
          in: path
          required: true
          description: The key of the experiment that should be provided
          schema:
            type: string
      responses:
        '200':
          description: An object containing the requested experiment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Merge/Patch a experiment object into the existing experiment object
      operationId: updateExperiment
      tags:
        - experiments
      parameters:
        - name: projectKey
          in: path
          required: true
          description: The key of the project that the run belongs to
          schema:
            type: string
        - name: experimentKey
          in: path
          required: true
          description: The key of the experiment that should be updated
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/Experiment"
      responses:
        '204':
          description: The operation was successful
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    ArtifactRef:
      type: object
      required:
        - name
        - version
      properties:
        name:
          type: string
        version:
          type: integer
          format: int32
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          example: 404
        message:
          type: string
          example: The requested resource does not exist
    Experiment:
      type: object
      required:
        - createdAt
        - key
        - name
        - status
      properties:
        createdAt:
          type: string
          format: date-time
          readOnly: true
        key:
          type: string
          example: linear-regression-with-sklearn
        name:
          type: string
          example: linear regression with sklearn
        tags:
          type: array
          items:
            type: string
          example:
            - sklearn
            - failed
        status:
          type: string
          enum:
            - TODO
            - IN_PROGRESS
            - COMPLETED
    Experiments:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Experiment"
...
    Run:
      type: object
      required:
        - id
        - status
        - createdBy
        - createdAt
      properties:
        artifacts:
          type: array
          readOnly: true
          items:
            $ref: "#/components/schemas/ArtifactRef"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          $ref: "#/components/schemas/UserRef"
        endTime:
          type: string
          format: date-time
          readOnly: true
        experimentRefs:
          type: array
          items:
            properties:
              experimentKey:
                type: string
        id:
          type: string
          readOnly: true
          example: ed9b4c00-d1d7-4dd4-b58b-008f5fc587a5
        metrics:
          $ref: "#/components/schemas/RunMetric"
        name:
          type: string
          example: training model with sklearn
        note:
          type: string
          example: should try better though
        parameters:
          $ref: "#/components/schemas/RunParameter"
        startTime:
          type: string
          format: date-time
          readOnly: true
        status:
          type: string
          enum:
            - RUNNING
            - COMPLETED
            - FAILED
          readOnly: true
        usedArtifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactRef"
    Runs:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Run"
    RunMetric:
      type: object
      additionalProperties: true
      example:
        acc: 98
        rmse: 98
    RunParameter:
      type: object
      additionalProperties: true
      example:
        batch_size: 512
        mse: 0.1
...
    UserRef:
      type: object
      readOnly: true
      required:
        - userId
        - nickName
      properties:
        userId:
          type: string
          readOnly: true
          example: 5def26a1-a932-404a-998c-aabf84994dd0
        nickName:
          type: string
          readOnly: true
          example: Alice
  securitySchemes:
    oauth2:
      type: oauth2
      description: This API uses OAuth 2 with the Authorization Code Flow with Proof Key for Code Exchange (PKCE).
      flows:
        authorizationCode:
          authorizationUrl: https://mlaide-dev.eu.auth0.com/authorize
          tokenUrl: https://mlaide-dev.eu.auth0.com/oauth/token
          scopes:
            projects:create: Create new projects
            projects:edit: Edit existing projects